Renames AboutPreferencesActivity to AboutActivity Moves it from package preferences to activities
